changed Item value to Float instead of Int.

+++ app/src/main/java/com/example/mapapp/domain/ChangeItemOpacityUseCase.kt
@@ -5,7 +5,7 @@ import com.example.mapapp.domain.entity.Item
 class ChangeItemOpacityUseCase(
     private val repository: Repository
 ) {
-    operator fun invoke(item: Item, value: Int) {
+    operator fun invoke(item: Item, value: Float) {
         repository.changeItemOpacity(item, value)
     }
 }
